+++ src/dal/remote/factory.py
@@ -1,6 +1,8 @@
 
 
 
+
+from src.dal.remote.producthunt_adapter import ProductHuntAdapter
 from src.dal.remote.devto_adapter import DevToAdapter
 from src.dal.remote.aws.aws_infra_catalog_adapter import  AwsInfraCatalogAdapter # AwsWhitepaperServicesAdapter
 from src.dal.remote.killedbygoogle_adapter import KilledByGoogleAdapter
@@ -24,6 +26,7 @@ class AdapterFactory:
         # "aws_whitepaper_services": AwsWhitepaperServicesAdapter # have a api that is better than scraping
         "aws_infra_catalog": AwsInfraCatalogAdapter,
         "devto": DevToAdapter,
+        "product_hunt": ProductHuntAdapter
     }
 
     @classmethod